requirements.txt
PyYAML
urllib3
tqdm
requests
pycaption
colorama
protobuf
pycryptodomex
pycountry
beautifulsoup4>=4.8.1
xmltodict